Download models directly into target/classes
It is best practice not to generate sources or resources into src/main,
but rather into target/generated-sources or similar.

In the case of a stable download, of course, it makes sense to cache the
download after the first time; fortunately, the download-maven-plugin
takes care of doing that for us:

  [INFO] --- download-maven-plugin:1.3.0:wget (download-inception-model) @ samples ---
  [INFO] Got from cache: /Users/curtis/.m2/repository/.cache/download-maven-plugin/inception5h.zip_176c69c52a422bb199f5543bf0537350

Hence, even though "mvn clean" wipes out the entire target folder,
it is still OK to download the model(s) directly into target/classes.