Add super-elegant Eclipse-specific configuration
Without this, importing the project into Eclipse yields the error
"Plugin execution not covered by lifecycle configuration" for the goal
com.googlecode.maven-download-plugin:download-maven-plugin:1.3.0:wget.

In this case, it is (relatively) simple to satisfy Eclipse: we declare
that this goal should run during the initial configuration step only,
and not incrementally.