LabelImage: use TensorFlowService to get the graph
That way, after loading it once, the graph is cached for later reuse.
However, there is not yet visual feedback on the download + unpacking.

We reference the remote model using a SciJava HTTPLocation, meaning we
no longer have to use the download-maven-plugin at build time to get it.
Instead, we download it on demand at runtime. This reduces POM config.