Improve the Tensors API
* No more normalization
* 3 kind of functions for every direction:
  * tensor(RAI) and img<Type>(Tensor): Reverses the order of dimensions
    but may use the storage array of an ArrayImg
  * tensor(RAI, int[]) and img<Type>(Tensor, int[]): Uses a costum
    dimension mapping.
  * tensorDirect(RAI) and img<Type>Direct(Tensor): Maps the dimensions
    1:1 between ImageJ and TensorFlow.
* For Tensor -> Img there are functions for every supported type,
  because Tensors are not typed but images are.

Why do we want all this functions:

The functions with the custom mapping will be probably used most often
because TensorFlow uses YXC for 2D images with channels by default and
this can not be done with the other methods. But I think this shouldn't
go into the API because there is no default dimension order for N-D
images (Or I didn't find it).

The reverse mapping function is useful because it may uses the
underlying array of an ArrayImg and it is fine for images without
multiple channels.

The direct function gives the user an easier way to handle dimensions
himself.

Naming:

The tensorsDirect and tensor function could switch their names but I
think the current naming is the most intuitve: "Direct" refers to the
dimensions.

Initial changes by: Curtis Rueden <ctrueden@wisc.edu>